Improved CustomItemEventHander

+++ UncomplicatedCustomItems/Example/ExampleCustomItem.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using Exiled.API.Enums;
 using Exiled.Events.EventArgs.Player;
 using LabApi.Events.Arguments.PlayerEvents;
+using UncomplicatedCustomItems.API;
 using UncomplicatedCustomItems.API.Attributes;
 using UncomplicatedCustomItems.API.Features;
 using UncomplicatedCustomItems.API.Features.SpecificData;
@@ -107,7 +108,8 @@ namespace UncomplicatedCustomItems.Examples
         // Register with CustomItemEventHandler.Init<CUSTOMITEMNAMESPACE.Events>();
         public override void OnShot(ShotEventArgs ev)
         {
-            ev.Player.Kill(ev.Firearm.HitscanHitregModule.GetHandler(ev.Damage));
+            if (Utilities.TryGetSummonedCustomItem(ev.Item.Serial, out var item) && item.CustomItem.Id == 1)
+                ev.Player.Kill(ev.Firearm.HitscanHitregModule.GetHandler(ev.Damage));
         }
     }
 }
